requirements.txt
coverage==5.3
infix==1.2
MIDIUtil==1.2.1
mido==1.2.10
pyfluidsynth==1.3.0
toolz==0.11.1
